Demystify Blockscout contract verification
`helpers.etherscan.verify` is our tooling wrapping `hardhat-verify`
plugin. We could name this tooling differently, but the `hardhat-verify`
plugin is configured with `etherscan` entry in the Hardhat config even
though it may be working with another block explorer, set via
`config.etherscan.customChains.urls`. We keep the tooling name like that
to stay consistent with the config of the plugin we are wrapping.

This commit attempts to demystify it by removing the etherscan tag from
the network config condition and adding a comment explaining it is
Blockscout verification under the hood.