Mezo Forwarder contract implementation
The contract code is copied from the GSN implementation which is the recommended
forwarder implementation for OpenZeppelin 4.x we use in the project.

The forwarder is a trusted contract allowing to perform smart contract
operations on behalf of another address with off-chain signatures. All contracts
in the codebase extending `ERC2771Context` and using `_msgSender()` instead of
`msg.sender` support calls via the forwarder.

My first reaction was to remove all those occurrences given Bitcoin wallet calls
are going to be relayed with the `msg.sender` being the Safe and for ETH
wallets, we do not want to sponsor calls. Then I realized, ERC-2771 could not
only be about sponsoring calls but also about off-chain precommitments to some
actions. For example, we could gather off-chain precommitments to specific
governance decisions, providing liquidity, etc, and have some relayer submit
them to the chain when the time comes. That said, given this is an audited code
it is probably better to leave it as-is and not close the doors for relaying
transactions, even for ETH wallets. Hence, the GSN forwarder!