Renaming FeeSplitter -> ChainFeeSplitter
We will have more than one splitter in mezodrome and we plan to abstract
some of the splitter logic into a new contract. ChainFeeSplitter is a
more concrete name for this functionality and at the same time
FeeSplitter or Splitter will be a more high level name for the abstract
contract.