Abstracting Fee Splitter functionality
At some point Mezodrome will have more than 1 splitter. A lot of the
ChainFeeSplitter code can be abstracted and moved to an abstract
Splitter contract. The common parts for splitters should be moving a
needle up or down. However, updating the period and distributing funds
might vary among the splitters, hence abstracting a signature only for
this functionality.