Abstracting nudge() and part of updatePeriod()
Turns out we can assume that all of the splitters will base on the same
TICK and WEEKly update interval. Having this in mind we can abstract
entire nudge() function and make only transfer part base on the concrete
implementation. In this case we make 2 virtual functions:
transferFirstRecipient and transferSecondRecipient that should be
implemented in the concrete splitter.