Remove ETH/WETH support from the Router
The default Router implementation exposes paths allowing to
add/remove liquidity and swap/zap using native ETH transfers.
Such ETH is converted to WETH back and forth under the hood.
This is a must because native ETH cannot be used in pools directly.

However, Mezo native token is BTC which is also exposed as an ERC20
precompiled contract. Native and ERC20 actions on BTC are equivalent
and unlike for ETH and WETH, there is no intermediary conversion step.

That said, a special support for native BTC transfers in the Router is
redundant as the BTC ERC20 precompile can be used directly in all cases.
Because of that, we decided to remove it from the Router and greatly
simplify the code. Moreover, if we kept the separate path for native
transfers, we would strongly complicate Router tests. This is because we
use a mock BTC implementation there. This mock doesn't have the
same native <-> ERC20 equivalency as the actual BTC token on Mezo chain
and this feature would be hard to mimic in a test environment.