Remove direct `Voter` dependency from the `Router`
The `voter` field in the `Router` is used in one specific case during `zapIn`, to fetch the pool gauge for direct staking. We don't want to deploy the `VeBTCVoter` and `VeBTC` machinery yet, hence, we make the `voter` dependency in the `Router` optional. The `deployer` can set it separately, later.