TokenGrant contract groundworks
Created TokenGrant contract with a draft of the basic logic allowing to
convert TokenGrant to a ve NFT. This will allow us to pursue the work
simultaneously on TokenGrant and what happens in ve NFT in regards to
merging, managed NFTs and revoking. Beware, it's just a scaffolding
allowing us to avoid merge conflicts. This code is not covered in tests
as of now.