Capture the vesting end and the grant manager in the veNFT
Implemented createGrantLockFor to create veNFT from TokenGrants. When
this path is used, grant manager and vesting end date are captured in
veNFT storage. Those two will be used to block/allow certain actions in
veNFT created from a grant. For example, we should allow revoking the
grant and we should disallow merging veNFTs if at least one of them is
an active grant.