Block merging veNFTs created for unvested grants
From now on, `merge` reverts if at least one of veNFTs being merged was
created from a grant and that grant has not fully vested yet.

`merge` allows creating one veNFT from two veNFTs. The source veNFT is
burned and underlying balance is moved from the source veNFT to the
target veNFT. When one of the veNFTs is a grant, the accounting gets
really complicated, so we disallow this scenario altogether.