Block splitting veNFT created for unvested grants
From now on, `split` reverts if the veNFT being split was created from a
grant and that grant has not fully vested yet.

`split` allows creating two veNFTs from one veNFT. In theory, if we
maintain the vesting end and grant manager for the two veNFTs split we
should be fine. In practice, we block merging unvested grant veNFTs so
we should be consistent in the other direction. If merging is not
possible, splitting would be irreversible action.