Add VestingWalletCliffUpgradeable contract
The VestingWalletCliffUpgradeable contract is a copy of the original file from
OpenZeppelin Contracts v5 with the following changes:
- The storage structure is aligned with the pattern used in OpenZeppelin Contracts v4:
  - Defined _cliff as a variable instead of storage struct.
- The `end` function is added to the contract.

This repository uses OpenZeppelin Contracts v4.9.0 and to avoid unexpected
breakage when upgrading to v5, we need to keep the contract in the same
format as in v4.

The contract is used in the TokenGrant contract to add a cliff to the vesting
schedule.