Link Escrow library with Grant library
The Grant library performs a call to _createLock function which is
defined in the Escrow library as an external function. Given it is a
delegatecall and the address of Escrow needs to be known, we need to
link those two libraries.

Alternative option would be to make _createLock internal but it adds
1.16KB to the token contract size and is not worth it.