Final implementation for TokenGrant.convert and unit tests
`TokenGrant.convert` converts a token grant to a veNFT with lock time
equal to the remaining vesting schedule duration rounded down to the
nearest week. The operation is irreversible and takes the entire token
balance of TokenGrant. The function fails if the token balance is zero.
Only grant beneficiary can perform the conversion. The function can be
called more than one time if the `TokenGrant` token balance increased
after the previous conversion. For each `convert()` call, a new veNFT is
created. The veNFT will use the same vesting schedule end, no matter
when it is called.