Define VestingWalletUpgradeable contract
The contract is copied from Open Zeppelin Contracts v5 with some adjustments,
from which the most significant is removing support of multiple tokens and
native asset. We want to simplify the solution, so we can enhance it
with revoke feature more easily.

The contract will be good to support multiple assets in different instances,
including Mezo's native Bitcoin gas asset, which is also IERC20.