Set the default needle in the fee splitter to 60:40
The `ChainFeeSplitter` contract should initially distribute 60% of the chain fees to the veBTC holders and another 40% to staking gauges. The business motivation behind this decision is that we will soon have veBTC holders that will look for yield but staking gauges is pushed more in the future. There is no sense to give more fees to gauges in that case hence we want to promote veBTC holders initially.