Expose `/metrics` endpoint for Prometheus
We want to be notified if something goes wrong so that we can take
action manually, so we need to integrate with our monitoring system.
Exposing the timestamp of the last successful transaction
(`updatePeriod`) allows us to detect potential issues. If no transaction
has been completed within the past week, it may indicate a problem with
maintainer that requires our attention.